feat: moved token utils into Earn directory
