+++ app/components/UI/Earn/utils/index.ts
@@ -1,1 +1,2 @@
+export * from './token';
 export * from './string';
